back/requirements.txt
fastapi
python-dotenv
neo4j
pydantic[email]
passlib
python-jose
python-multipart
starlette
uvicorn
bcrypt
httpx